# docs/swagger.yaml
definitions:
  admin_dto.Album:
    properties:
      created_at:
        type: integer
      id:
        type: string
      participant_count:
        type: integer
      title:
        type: string
      total_cards:
        type: integer
    type: object
  admin_dto.CreateCardRequest:
    properties:
      album_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
  admin_dto.Message:
    properties:
      message:
        type: string
    type: object
  admin_dto.Overview:
    properties:
      albums:
        type: integer
      cards:
        type: integer
      collected:
        type: integer
      contacts:
        type: integer
      participants:
        type: integer
      users:
        type: integer
    type: object
  admin_dto.UpdateAlbumRequest:
    properties:
      title:
        type: string
    type: object
  admin_dto.UpdateCardRequest:
    properties:
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
  admin_dto.UpdateUserRoleRequest:
    properties:
      is_admin:
        type: boolean
    type: object
  admin_dto.User:
    properties:
      created_at:
        type: integer
      description:
        type: string
      email:
        type: string
      github:
        type: string
      id:
        type: string
      is_admin:
        type: integer
      linkedin:
        type: string
      name:
        type: string
      picture:
        type: string
      updated_at:
        type: integer
      web:
        type: string
    type: object
  admin_dto.UserDetail:
    properties:
      albums:
        items:
          $ref: '#/definitions/admin_dto.Album'
        type: array
      cards:
        items:
          $ref: '#/definitions/card_model.Card'
        type: array
      user:
        $ref: '#/definitions/admin_dto.User'
    type: object
  album_dto.CreateAlbumCardRequest:
    properties:
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
  album_dto.CreateAlbumRequest:
    properties:
      cards:
        items:
          $ref: '#/definitions/album_dto.CreateAlbumCardRequest'
        type: array
      title:
        type: string
    type: object
...
  album_model.Album:
    properties:
      cards:
        items:
          $ref: '#/definitions/card_model.Card'
        type: array
      created_at:
        type: integer
      id:
        type: string
      title:
        type: string
      total_cards:
        type: integer
    type: object
...
  card_model.Card:
    properties:
      album_id:
        type: string
      created_at:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
      updated_at:
        type: integer
    type: object
...
info:
  contact: {}
  description: MVP para intercambio de tarjetas coleccionables.
  title: Trading Card Album API
  version: "1.0"
paths:
  /api/v1/admin/albums:
    get:
      consumes:
      - application/json
      description: List all albums with stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_dto.Album'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create an album with its cards
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/album_dto.CreateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/album_model.Album'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/albums/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Message'
      security:
      - BearerAuth: []
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update album title
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.UpdateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Album'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/albums/{id}/cards:
    get:
      consumes:
      - application/json
      description: Get cards of an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/card_model.Card'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/cards:
    get:
      consumes:
      - application/json
      description: List all cards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/card_model.Card'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a card in an album
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.CreateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/cards/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Message'
      security:
      - BearerAuth: []
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update a card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/overview:
    get:
      consumes:
      - application/json
      description: Get platform overview metrics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Overview'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/users:
    get:
      consumes:
      - application/json
      description: List users, optionally filtered by email
      parameters:
      - description: Filter by email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_dto.User'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Get user detail with albums and cards
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.UserDetail'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Update user admin role
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.User'
      security:
      - BearerAuth: []
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"